surgery/requirements.txt
# Phase 2: Live Intra-Op AR Pipeline
# GPU-oriented install for Linux with NVIDIA hardware.

holoscan>=4.1.0

torch>=2.5.1
torchvision>=0.20.1

MedSAM2 @ git+https://github.com/bowang-lab/MedSAM2.git

ultralytics>=8.1.0
inference-sdk>=0.20.0
roboflow>=1.1.60

opencv-python>=4.8.0
numpy>=1.24.0
PyYAML>=6.0
requests>=2.32.0
hydra-core>=1.3.2
SimpleITK>=2.4.0
tqdm>=4.66.5
streamlit>=1.44.0

pytest>=8.0
